feat: add custom title for pages
The title of a page displays the most important information, mainly its
hierarchical position on the website.

For most of the pages this is just the hierarchical position, e.g. for
route '/archivedprojects' the title is 'Archived Projects'.
For all pages, which belong to a project, '| {data.project.name}' is
appended.
In the paper view, the paper title is the title of the page. This way,
it's easy to differentiate pages of different papers.
In general everything is appended with '| SnowballR'.