feat: display submitted review when navigating to the view of a reviewed paper
The submitted review is shown only when the review mode is activated, otherwise all review information are shown. Furthermore, all checkboxes and buttons are disabled after a review was submitted at the moment indicating the editing of a review is not possible yet. The page data are not invalidated and refetched as it will be implemented soon that the next paper will be automatically opened after submitting the review