refactor: make tab list trigger a tag when href is provided
Previously, we an a tag with a nested button, but the button is never clicked.
Changing the logic of the tabs trigger so that it switches to an a tag when href is provided
fixes this as now only the a tag is shown, but when no href is provided a button is rendered instead.