refactor: handle selecting and unselecting of all options in 'Select' component internally different
Previously, the 'Select all / Unselect all' select item could only be clicked. With this change, it behaves like a normal select and can be also selected / unselected via pressing space or enter. The only difference to the other select items is that the no checkmark is displayed in front of the label, but this label changes to indicate that the option was selected.