fix: wrong focus to suggestion list instead of suggestion items in the 'InviteUsersInput' component
Previously, the focus was set to the list element which works in Safari but not Firefox or Chrome. To manage the focus correctly in all browsers we only select the suggestions and scroll the selected suggestion into the view and keep the focus on the input, which also has the advantage that the text can still be changed.